Update docker workflow to support multiple environments
Updates docker.yml to be a generic, reusable workflow with inputs
to specify the hugo environment to build and the name of the
container image to create.

Adds new workflows for the production and staging environments
which delegate to the reusable docker.yml workflow with inputs
appropriate for each environment.